+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Disable UI Navigation when quitting the game or loading a new Scene
+
 ## [0.3.0] - 2025-10-06
 ### Added
 - MenuController